moved unit test to the right file
