+++ tests/core_tests/test_template.py
@@ -8,7 +8,7 @@
 # agreement to the Shotgun Pipeline Toolkit Source Code License. All rights 
 # not expressly granted therein are reserved by Shotgun Software Inc.
 
-from __future__ import with_statement
+import sys
 import os
 
 import tank
@@ -378,222 +378,3 @@ class TestReadTemplates(TankTestBase):
         self.assertEquals(["Seq", "Shot"], key.exclusions)
 
 
-class TestIntegerKey(TankTestBase):
-
-    def test_init_validation(self):
-        """
-        Makes sure that parameter validation is correct in the constructor.
-        """
-        # This should obviously work
-        self._validate_key(
-            IntegerKey("version_number"),
-            strict_matching=None, format_spec=None
-        )
-        # When specifying parameters, they should be set accordingly.
-        self._validate_key(
-            IntegerKey("version_number", format_spec="03", strict_matching=True),
-            strict_matching=True, format_spec="03"
-        )
-        self._validate_key(
-            IntegerKey("version_number", format_spec="03", strict_matching=False),
-            strict_matching=False, format_spec="03"
-        )
-        self._validate_key(
-            IntegerKey("version_number", format_spec="3", strict_matching=False),
-            strict_matching=False, format_spec="3"
-        )
-        # When specifying a format but not specifying the strict_matching, it should
-        # still have strict_matching.
-        self._validate_key(
-            IntegerKey("version_number", format_spec="03"),
-            strict_matching=True, format_spec="03"
-        )
-        # Make sure than an error is raised when wrong types are passed in.
-        with self.assertRaisesRegexp(TankError, "is not of type boolean"):
-            IntegerKey("version_number", strict_matching=1)
-
-        with self.assertRaisesRegexp(TankError, "is not of type string"):
-            IntegerKey("version_number", format_spec=1)
-
-        # Make sure that if the user specifies strict_matching with no format
-        # there is an error
-        error_regexp = "strict_matching can't be set"
-        with self.assertRaisesRegexp(TankError, error_regexp):
-            IntegerKey("version_number", strict_matching=False)
-
-        with self.assertRaisesRegexp(TankError, error_regexp):
-            IntegerKey("version_number", strict_matching=True)
-
-        # We support 4 format_spec values:
-        # - None
-        # - non zero positive number
-        # - zero followed by a non zero positive number
-        IntegerKey("version_number", format_spec=None)
-        IntegerKey("version_number", format_spec="1")
-        IntegerKey("version_number", format_spec="01")
-
-        # Make sure invalid formats are caught
-        with self.assertRaisesRegexp(TankError, "format_spec can't be empty"):
-            IntegerKey("version_number", format_spec="")
-
-        error_regexp = "has to either be"
-        # We don't support the sign option.
-        with self.assertRaisesRegexp(TankError, error_regexp):
-            IntegerKey("version_number", format_spec=" 3", strict_matching=False)
-
-        with self.assertRaisesRegexp(TankError, error_regexp):
-            # Should throw because the padding number is not non zero.
-            IntegerKey("version_number", format_spec="00")
-
-        with self.assertRaisesRegexp(TankError, error_regexp):
-            # Should throw because it is not a non zero positive integer
-            IntegerKey("version_number", format_spec="0")
-
-        with self.assertRaisesRegexp(TankError, error_regexp):
-            # Should throw because the padding caracter is invalid
-            IntegerKey("version_number", format_spec="a0")
-
-        with self.assertRaisesRegexp(TankError, error_regexp):
-            # Should throw because the padding size is not a number.
-            IntegerKey("version_number", format_spec="0a")
-
-    def test_no_format_spec(self):
-        key = IntegerKey("version_number")
-        key.value_from_str
-
-    def _validate_key(self, key, strict_matching, format_spec):
-        """
-        Makes sure that an integer key's formatting options are correctly set.
-        """
-        self.assertEqual(key.strict_matching, strict_matching)
-        self.assertEqual(key.format_spec, format_spec)
-
-    def test_non_strict_matching(self):
-        """
-        In non strict mode, tokens can actually have less numbers than the padding requests. Also,
-        if there are more, can also match.
-        """
-        self._test_non_strict_matching('0')
-        self._test_non_strict_matching('')
-
-    def _test_non_strict_matching(self, padding_char):
-        """
-        Allows to test strict matching with a specific padding character.
-
-        :param padding_char: Character to test padding with. Should be space or 0.
-        """
-        key = IntegerKey("version_number", format_spec="%s3" % padding_char, strict_matching=False)
-        keys = {
-            "version_number": key
-        }
-        definition = "v_{version_number}"
-        tpl = TemplatePath(definition, keys, "")
-
-        # The padding char is missing in the format specifier, but we still need when validating
-        # results.
-        if padding_char == '':
-            padding_char = ' '
-
-        # It should match because they are valid numbers.
-        fields = tpl.validate_and_get_fields("v_000")
-        self.assertEqual(fields, {"version_number": 0})
-        fields = tpl.validate_and_get_fields("v_00")
-        self.assertEqual(fields, {"version_number": 0})
-        fields = tpl.validate_and_get_fields("v_0")
-        self.assertEqual(fields, {"version_number": 0})
-
-        # While the number doesn't make any sense as far as formatting is concerned, this used
-        # to work in old versions of Toolkit and needs to keep working in non strict mode.
-        fields = tpl.validate_and_get_fields("v_%s000" % padding_char)
-        self.assertEqual(fields, {"version_number": 0})
-
-        # It should match a template with too many digits...
-        fields = tpl.validate_and_get_fields("v_20000")
-        self.assertEqual(fields, {"version_number": 20000})
-
-        # ... even if they are all zeros because lossy matching.
-        fields = tpl.validate_and_get_fields("v_00000")
-        self.assertEqual(fields, {"version_number": 0})
-
-        # From path to tokens back to path should be lossy.
-        fields = tpl.validate_and_get_fields("v_1")
-        self.assertEqual("v_%s%s1" % (padding_char, padding_char), tpl.apply_fields(fields))
-
-        self._test_nan(key, "expected an Integer")
-
-    def _test_nan(self, key, error_msg):
-        """
-        Tests a key with against values that are not numbers.
-
-        :param key: Key to test.
-        :param error_msg: Text that partially matches the error message.
-        """
-        # Should fail because not a number
-        with self.assertRaisesRegexp(TankError, error_msg):
-            key.value_from_str("aaaa")
-        with self.assertRaisesRegexp(TankError, error_msg):
-            key.value_from_str("0a")
-        with self.assertRaisesRegexp(TankError, error_msg):
-            key.value_from_str("a0")
-        with self.assertRaisesRegexp(TankError, error_msg):
-            key.value_from_str("aa")
-
-    def _test_strict_matching(self, padding_char):
-        """
-        Allows to test strict matching with any padding type.
-        """
-        # have a template that formats with two digits of padding.
-        key = IntegerKey("version_number", format_spec="%s3" % padding_char, strict_matching=True)
-        keys = {
-            "version_number": key
-        }
-        self.assertTrue(keys["version_number"].strict_matching)
-        definition = "v{version_number}"
-        tpl = TemplatePath(definition, keys, "")
-
-        # The padding char is missing in the format specifier, but we still need when validating
-        # results.
-        if padding_char == '':
-            padding_char = ' '
-
-        # It should not match a template with only one digit.
-        fields = tpl.validate_and_get_fields("v1")
-        self.assertIsNone(fields)
-
-        # From path to tokens back to path should get back the same string when the expected number of
-        # digits are found.
-        fields = tpl.validate_and_get_fields("v%s%s1" % (padding_char, padding_char))
-        self.assertEqual("v%s%s1" % (padding_char, padding_char), tpl.apply_fields(fields))
-
-        # It should match a template with too many digits.
-        fields = tpl.validate_and_get_fields("v20000")
-        self.assertEqual(fields, {"version_number": 20000})
-        self.assertEqual("v20000", tpl.apply_fields(fields))
-
-        error_msg = "does not match format spec"
-        # It should not match a template with too many digits that are all zero because that would
-        # lossy. (there are more zeros than the format spec can rebuild)
-        with self.assertRaisesRegexp(TankError, error_msg):
-            key.value_from_str("0000")
-
-        # It should not match negative numbers either
-        with self.assertRaisesRegexp(TankError, error_msg):
-            key.value_from_str("-1000")
-
-        # It should not match baddly padded numbers
-        with self.assertRaisesRegexp(TankError, error_msg):
-            key.value_from_str("0100")
-
-        # It should not match negative values
-        with self.assertRaisesRegexp(TankError, error_msg):
-            key.value_from_str("-01")
-
-        self._test_nan(key, error_msg)
-
-    def test_strict_matching(self):
-        """
-        In strict mode, tokens have to have as much padding as the format specifier suggests. Less will not
-        match.
-        """
-        self._test_strict_matching('')
-        self._test_strict_matching('0')
